+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu Product Bundle
 
 * dev-develop
 
+    * FEATURE     Added functionality for adding product and product with variants as addons.
     * ENHANCEMENT Not showing variants in product list.
     * FEATURE     Changed delete behavior when product contains variants.
     * FEATURE     Added UI for handling variants.